+++ mpitests/test_expval.py
@@ -31,8 +31,6 @@ class TestExpval:
     def test_identity_expectation(self, theta, phi, tol):
         """Test that identity expectation value (i.e. the trace) is 1"""
         dev = qml.device(device_name, mpi=True, wires=3)
-        if device_name == "lightning.gpu" and dev.R_DTYPE == np.float32:
-            pytest.skip("Skipped FP32 tests for expval in lightning.gpu")
 
         O1 = qml.Identity(wires=[0])
         O2 = qml.Identity(wires=[1])
@@ -49,9 +47,6 @@ class TestExpval:
         """Test that PauliZ expectation value is correct"""
         dev = qml.device(device_name, mpi=True, wires=3)
 
-        if device_name == "lightning.gpu" and dev.R_DTYPE == np.float32:
-            pytest.skip("Skipped FP32 tests for expval in lightning.gpu")
-
         O1 = qml.PauliZ(wires=[0])
         O2 = qml.PauliZ(wires=[1])
 
@@ -66,9 +61,6 @@ class TestExpval:
     def test_paulix_expectation(self, theta, phi, tol):
         """Test that PauliX expectation value is correct"""
         dev = qml.device(device_name, mpi=True, wires=3)
-
-        if device_name == "lightning.gpu" and dev.R_DTYPE == np.float32:
-            pytest.skip("Skipped FP32 tests for expval in lightning.gpu")
 
         O1 = qml.PauliX(wires=[0])
         O2 = qml.PauliX(wires=[1])
@@ -88,9 +80,6 @@ class TestExpval:
     def test_pauliy_expectation(self, theta, phi, tol):
         """Test that PauliY expectation value is correct"""
         dev = qml.device(device_name, mpi=True, wires=3)
-
-        if device_name == "lightning.gpu" and dev.R_DTYPE == np.float32:
-            pytest.skip("Skipped FP32 tests for expval in lightning.gpu")
 
         O1 = qml.PauliY(wires=[0])
         O2 = qml.PauliY(wires=[1])
@@ -130,8 +119,6 @@ class TestExpval:
         n_qubits = 7
         dev_def = qml.device("default.qubit", wires=n_qubits)
         dev = qml.device(device_name, mpi=True, wires=n_qubits)
-        if device_name == "lightning.gpu" and dev.R_DTYPE == np.float32:
-            pytest.skip("Skipped FP32 tests for expval in lightning.gpu")
         comm = MPI.COMM_WORLD
 
         m = 2**n_wires