+++ pennylane_lightning/core/src/simulators/lightning_gpu/StateVectorCudaMPI.hpp
@@ -38,6 +38,8 @@
 #include "cuGates_host.hpp"
 #include "cuda_helpers.hpp"
 
+#include "CPUMemoryModel.hpp"
+
 #include "LinearAlg.hpp"
 
 /// @cond DEV
@@ -97,6 +99,7 @@ class StateVectorCudaMPI final
                                      StateVectorCudaMPI<Precision>>::CFP_t;
     using PrecisionT = Precision;
     using ComplexT = std::complex<PrecisionT>;
+    using MemoryStorageT = Pennylane::Util::MemoryStorageLocation::Undefined;
 
     StateVectorCudaMPI() = delete;
 
@@ -204,6 +207,8 @@ class StateVectorCudaMPI final
               handle_.get(), mpi_manager_, 0, BaseType::getData(),
               numLocalQubits_, localStream_.get())),
           gate_cache_(true, other.getDataBuffer().getDevTag()) {
+        PL_CUDA_IS_SUCCESS(cudaDeviceSynchronize());
+        mpi_manager_.Barrier();
         BaseType::CopyGpuDataToGpuIn(other.getData(), other.getLength(), false);
         PL_CUDA_IS_SUCCESS(cudaDeviceSynchronize());
         mpi_manager_.Barrier();
@@ -327,6 +332,31 @@ class StateVectorCudaMPI final
                             thread_per_block, stream_id);
         PL_CUDA_IS_SUCCESS(cudaDeviceSynchronize());
         mpi_manager_.Barrier();
+    }
+
+    /**
+     * @brief Apply a single gate to the state-vector. Offloads to custatevec
+     * specific API calls if available. If unable, attempts to use prior cached
+     * gate values on the device. Lastly, accepts a host-provided matrix if
+     * otherwise, and caches on the device for later reuse.
+     *
+     * @param opName Name of gate to apply.
+     * @param wires Wires to apply gate to.
+     * @param adjoint Indicates whether to use adjoint of gate.
+     * @param params Optional parameter list for parametric gates.
+     * @param matrix Matrix representation of gate.
+     */
+    void applyOperation(const std::string &opName,
+                        const std::vector<size_t> &wires, bool adjoint,
+                        const std::vector<Precision> &params,
+                        [[maybe_unused]] const std::vector<ComplexT> &matrix) {
+        std::vector<CFP_t> matrix_cu(matrix.size());
+        std::transform(matrix.begin(), matrix.end(), matrix_cu.begin(),
+                       [](const std::complex<Precision> &x) {
+                           return cuUtil::complexToCu<std::complex<Precision>>(
+                               x);
+                       });
+        applyOperation(opName, wires, adjoint, params, matrix_cu);
     }
 
     /**