+++ pennylane_lightning/core/src/simulators/lightning_gpu/measurements/MeasurementsGPUMPI.hpp
@@ -87,6 +87,7 @@ class MeasurementsMPI final
         } else {
             data_type_ = CUDA_C_32F;
         }
+        mpi_manager_.Barrier();
     };
 
     /**
@@ -209,6 +210,47 @@ class MeasurementsMPI final
     }
 
     /**
+     * @brief Probabilities to measure rotated basis states.
+     *
+     * @param obs An observable object.
+     * @param num_shots Number of shots(Optional).  If specified with a non-zero
+     * number, shot-noise will be added to return probabilities
+     *
+     * @return Floating point std::vector with probabilities
+     * in lexicographic order.
+     */
+    std::vector<PrecisionT> probs(const Observable<StateVectorT> &obs,
+                                  size_t num_shots = 0) {
+        return BaseType::probs(obs, num_shots);
+    }
+
+    /**
+     * @brief Probabilities with shot-noise.
+     *
+     * @param num_shots Number of shots.
+     *
+     * @return Floating point std::vector with probabilities.
+     */
+    std::vector<PrecisionT> probs(size_t num_shots) {
+        return BaseType::probs(num_shots);
+    }
+
+    /**
+     * @brief Probabilities with shot-noise for a subset of the full system.
+     *
+     * @param num_shots Number of shots.
+     * @param wires Wires will restrict probabilities to a subset
+     * of the full system.
+     *
+     * @return Floating point std::vector with probabilities.
+     */
+
+    std::vector<PrecisionT> probs(const std::vector<size_t> &wires,
+                                  size_t num_shots) {
+        return BaseType::probs(wires, num_shots);
+    }
+
+    /**
      * @brief Utility method for samples.
      *
      * @param num_samples Number of Samples
@@ -291,6 +333,7 @@ class MeasurementsMPI final
             precumulative = 0;
         }
         PL_CUDA_IS_SUCCESS(cudaDeviceSynchronize());
+        mpi_manager_.Barrier();
 
         // Ensure the 'custatevecSamplerApplySubSVOffset' function can be called
         // successfully without reducing accuracy.
@@ -320,6 +363,7 @@ class MeasurementsMPI final
         if (mpi_manager_.getRank() == 0) {
             preshotOffset = 0;
         }
+        mpi_manager_.Barrier();
 
         int nSubShots = shotOffset - preshotOffset;
         if (nSubShots > 0) {
@@ -341,6 +385,9 @@ class MeasurementsMPI final
             PL_CUDA_IS_SUCCESS(cudaFree(extraWorkspace));
         }
 
+        PL_CUDA_IS_SUCCESS(cudaDeviceSynchronize());
+        mpi_manager_.Barrier();
+
         mpi_manager_.Allreduce<custatevecIndex_t>(localBitStrings,
                                                   globalBitStrings, "sum");
 
@@ -350,6 +397,7 @@ class MeasurementsMPI final
                     (globalBitStrings[i] >> j) & 1U;
             }
         }
+        mpi_manager_.Barrier();
         return samples;
     }
 
@@ -484,6 +532,24 @@ class MeasurementsMPI final
         auto expect = mpi_manager_.allreduce<PrecisionT>(local_expect, "sum");
 
         return static_cast<PrecisionT>(expect);
+    }
+
+    /**
+     * @brief Expectation value for a Observable with shots
+     *
+     * @param obs Observable.
+     * @param num_shots Number of shots used to generate samples.
+     * @param shot_range The range of samples to use. All samples are used
+     * by default.
+     * @return Floating point expected value of the observable.
+     */
+
+    auto expval(const Observable<StateVectorT> &obs, const size_t &num_shots,
+                const std::vector<size_t> &shot_range) -> PrecisionT {
+        mpi_manager_.Barrier();
+        PrecisionT result = BaseType::expval(obs, num_shots, shot_range);
+        mpi_manager_.Barrier();
+        return result;
     }
 
     /**
@@ -709,5 +775,19 @@ class MeasurementsMPI final
 
         return mean_square - squared_mean * squared_mean;
     };
+
+    /**
+     * @brief Calculate the variance for an observable with the number of shots.
+     *
+     * @param obs An observable object.
+     * @param num_shots Number of shots.
+     *
+     * @return Variance of the given observable.
+     */
+
+    auto var(const Observable<StateVectorT> &obs, const size_t &num_shots)
+        -> PrecisionT {
+        return BaseType::var(obs, num_shots);
+    }
 }; // class Measurements
 } // namespace Pennylane::LightningGPU::Measures