+++ tests/test_serialize_chunk_obs.py
@@ -37,12 +37,18 @@ class TestSerializeObs:
     def test_chunk_obs(self, use_csingle, obs_chunk):
         """Test chunking of observable array"""
         with qml.tape.QuantumTape() as tape:
+            qml.expval(
+                0.5 * qml.PauliX(0) @ qml.PauliZ(1)
+                + 0.7 * qml.PauliZ(0) @ qml.PauliX(1)
+                + 1.2 * qml.PauliY(0) @ qml.PauliY(1)
+            )
             qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
             qml.expval(qml.PauliY(wires=1))
             qml.expval(qml.PauliX(0) @ qml.Hermitian([[0, 1], [1, 0]], wires=3) @ qml.Hadamard(2))
             qml.expval(qml.Hermitian(qml.PauliZ.compute_matrix(), wires=0) @ qml.Identity(1))
-        s = QuantumScriptSerializer(device_name, use_csingle).serialize_observables(
-            tape, self.wires_dict
-        )
+        s, offsets = QuantumScriptSerializer(
+            device_name, use_csingle, split_obs=True
+        ).serialize_observables(tape, self.wires_dict)
         obtained_chunks = pennylane_lightning.core.lightning_base._chunk_iterable(s, obs_chunk)
         assert len(list(obtained_chunks)) == int(np.ceil(len(s) / obs_chunk))
+        assert [0, 3, 4, 5, 6, 7] == offsets